Explicitly pin dependencies
This changes the dependency pinning to be explicit where it was
optimistic or pessimistic so that versions that were known to work with
the app are installed. This is an effort to get the app working.

A major thing to note is that the node version is set to v0.10.44, which
is the latest in the v0.10 series, which is what this app was built
with. Using a newer node causes the older dependencies to fail and not
install.