Use Firebase for hosting
This replaces Heroku with Firebase as the host since it works for
deploying and handles it as a static site. This is preferable to Heroku
because it is specifically for serving static sites, and it actually
works unlike Heroku.

Changes:

- `scripts/deploy.sh` can be used for deploying
- The `firebase.json` config is specifically for the app and Ember
- The Firebase used is the same for the data store
- The API has been updated to use CORS requests to the Sinatra app and
  uses a different URL for development and production

The issue with hosting on Heroku is that the app does not deploy and
work with the latest versions of the buildpack and stack.